pyproject.toml
[tool.poetry]
name = "compendium-backend"
authors = ["Yi Cao <me@ycao.top>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.37.1"
python-dotenv = "^1.0.1"
fastapi = "^0.111.1"
pillow = "^10.4.0"
uvicorn = {extras = ["standard"], version = "^0.30.3"}

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
